requirements.txt
databases==0.9.0
fastapi==0.115.8
passlib==1.7.4
pydantic==2.10.6
PyJWT==2.8.0
PyJWT==2.10.1
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.32.3
SQLAlchemy==2.0.34
uvicorn==0.34.0
